+++ dynamic_preferences/users/serializers.py
@@ -1,0 +1,5 @@
+from dynamic_preferences.api.serializers import PreferenceSerializer
+
+
+class UserPreferenceSerializer(PreferenceSerializer):
+    pass